requirements.txt
django
django-bootstrap3
django_fsu
jinja2
pygments
